% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeCrossover.R
\name{AnalyzeCrossover}
\alias{AnalyzeCrossover}
\title{Equivalence-Based Tests for Crossover Studies}
\usage{
AnalyzeCrossover(
  x = NULL,
  y = NULL,
  type = "welch",
  deltaL = -Inf,
  deltaU = Inf,
  alpha = NULL
)
}
\arguments{
\item{x}{a vector of data values with at least two distinct observations from Group 1. The observations should be the intra-subject \emph{differences} in the first sequence/group (i.e., of the form \eqn{x_{i} = x_{i1} - x_{i2}}, where \eqn{i} denotes the \eqn{i}th subject in the 1st sequence and the second subscript denotes the period). Please exclude missing data and \code{NA} values.}

\item{y}{a vector of data values with at least two distinct observations from Group 2. The observations should be the intra-subject \emph{differences} in the second sequence/group (i.e., of the form \eqn{y_{i} = y_{i1} - y_{i2}}, where \eqn{i} denotes the \eqn{i}th subject in the 2nd sequence and the second subscript denotes the period). Please exclude missing data and \code{NA} values.}

\item{type}{a character string specifying the type of \emph{t}-test(s) to be used, must be one of \code{"welch"} (default, does not assume intra-subject variance is the same in both groups) or \code{"student"} (assumes intra-subject variance is the same in both groups).}

\item{deltaL}{the lower bound for the interval of equivalence (can be \code{-Inf} for noninferiority test).}

\item{deltaU}{the upper bound for the interval of equivalence (can be \code{Inf} for noninferiority test).}

\item{alpha}{the significance level \eqn{\alpha \in (0,1)}.}
}
\value{
The test results are returned as a list with supplementary information. Equivalence or noninferiority is concluded when all \emph{p}-values in the Results table are less than \code{alpha}. Otherwise, we do not reject the null hypotheses of inequivalence or inferiority.
}
\description{
Conducts equivalence and noninferiority tests with two samples (including those with paired data).
}
\examples{
# specify targetPower to obtain sample sizes n
AnalyzeCrossover(x = rnorm(15), y = rnorm(15), type = "welch", deltaL = -1,
deltaU = 1, alpha = 0.05)

}
