% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOneSample.R
\name{AnalyzeOneSample}
\alias{AnalyzeOneSample}
\title{Equivalence-Based Tests for One-Group Studies}
\usage{
AnalyzeOneSample(x = NULL, deltaL = -Inf, deltaU = Inf, alpha = NULL)
}
\arguments{
\item{x}{a vector of data values with at least two distinct observations. Please exclude missing data and \code{NA} values.}

\item{deltaL}{the lower bound for the interval of equivalence (can be \code{-Inf} for nonsuperiority test).}

\item{deltaU}{the upper bound for the interval of equivalence (can be \code{Inf} for noninferiority test).}

\item{alpha}{the significance level \eqn{\alpha \in (0,1)}.}
}
\value{
The test results are returned as a list with supplementary information. Equivalence, noninferiority, or nonsuperiority is concluded when all \emph{p}-values in the Results table are less than \code{alpha}. Otherwise, we do not reject the null hypotheses of inequivalence, inferiority, or superiority.
}
\description{
Conducts equivalence and noninferiority tests with one sample.
}
\examples{
# analyze data from one group
AnalyzeOneSample(x = rnorm(15), deltaL = -1, deltaU = 1, alpha = 0.05)

}
