% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTwoSample.R
\name{AnalyzeTwoSample}
\alias{AnalyzeTwoSample}
\title{Equivalence-Based Tests for Two-Group Studies}
\usage{
AnalyzeTwoSample(
  x = NULL,
  y = NULL,
  type = "welch",
  deltaL = -Inf,
  deltaU = Inf,
  alpha = NULL
)
}
\arguments{
\item{x}{a vector of data values with at least two distinct observations from Group 1. Please exclude missing data and \code{NA} values.}

\item{y}{a vector of data values with at least two distinct observations from Group 2. Please exclude missing data and \code{NA} values. \code{x} and \code{y} must contain the same number of observations if \code{type = "paired"}.}

\item{type}{a character string specifying the type of \emph{t}-test(s) to be used, must be one of \code{"welch"} (default, for independent samples without equal variance assumption), \code{"student"} (independent sample with equal variance assumption), or \code{"paired"} (for paired data).}

\item{deltaL}{the lower bound for the interval of equivalence (can be \code{-Inf} for noninferiority test).}

\item{deltaU}{the upper bound for the interval of equivalence (can be \code{Inf} for noninferiority test).}

\item{alpha}{the significance level \eqn{\alpha \in (0,1)}.}
}
\value{
The test results are returned as a list with supplementary information. Equivalence or noninferiority is concluded when all \emph{p}-values in the Results table are less than \code{alpha}. Otherwise, we do not reject the null hypotheses of inequivalence or inferiority.
}
\description{
Conducts equivalence and noninferiority tests with two samples (including those with paired data).
}
\examples{
# analyze data from two groups
AnalyzeTwoSample(x = rnorm(15), y = rnorm(15), type = "welch", deltaL = -1,
deltaU = 1, alpha = 0.05)

}
