% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOneSample.R
\name{en.test-methods}
\alias{en.test-methods}
\alias{print.en.test}
\title{Printing equivalence and noninferiority test results}
\usage{
\method{print}{en.test}(x, ...)
}
\arguments{
\item{x}{an en.test object}

\item{...}{further arguments to be passed through (see \code{print()} function)}
}
\description{
Helper function to print en.test objects
}
\keyword{internal}
