% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designParallelUnequal.R
\name{power.en.test-methods}
\alias{power.en.test-methods}
\alias{print.power.en.test}
\title{Printing power analysis for equivalence and noninferiority tests}
\usage{
\method{print}{power.en.test}(x, ...)
}
\arguments{
\item{x}{an power.en.test object}

\item{...}{further arguments to be passed through (see \code{print()} function)}
}
\description{
Helper function to print power.en.test objects
}
\keyword{internal}
