% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designParallelUnequal.R
\name{UpdateTargetPower}
\alias{UpdateTargetPower}
\title{Updated Power Calculations for Equivalence and Noninferiority Tests}
\usage{
UpdateTargetPower(power.en.test = NULL, targetPower = NULL, plot = TRUE)
}
\arguments{
\item{power.en.test}{an object to which the output from a \code{Design...()} function call was saved. The input \code{targetPower} must have been specified in this function call so that the power curve was approximated. This function will return an error if the original \code{Design...()} function call specified \code{n}, \code{n1}, or \code{n2}. All other inputs from the initial \code{Design()} function call will be passed to this function.}

\item{targetPower}{the desired statistical power of the equivalence or noninferiority test, must be a single number between 0 and 1 (exclusive).}

\item{plot}{a logical variable indicating whether to return a plot of the power curve.}
}
\value{
The sample sizes or power estimate are returned as a list with supplementary information. A plot of the approximated power curve will appear on the plot pane if \code{plot = TRUE}.
}
\description{
Performs sample size calculations for equivalence and noninferiority tests using output saved from a previous function call with an updated target for statistical power.
}
\examples{
# save output from `Design...()` function call with previous `targetPower` specification
# to object `design.result`.
design.result <- DesignParallelUnequal(diff = -4, sigma1 = 15, sigma2 = 18, deltaL = -19.2,
                 deltaU = 19.2, alpha = 0.05, targetPower = 0.8, q = 1, seed = 1, sobol = 0)

UpdateTargetPower(power.en.test = design.result, targetPower = 0.7, plot = TRUE)

}
