% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designCrossoverUnequal.R
\name{DesignCrossover2x2Unequal}
\alias{DesignCrossover2x2Unequal}
\title{Power Calculations for Two-Group 2\eqn{\times}2 Crossover Designs with Unequal Variances}
\usage{
DesignCrossover2x2Unequal(
  diff = NULL,
  sigma1 = NULL,
  sigma2 = NULL,
  deltaL = -Inf,
  deltaU = Inf,
  alpha = NULL,
  targetPower = NULL,
  q = 1,
  n1 = NULL,
  n2 = NULL,
  plot = TRUE,
  seed = NULL,
  sobol = 0
)
}
\arguments{
\item{diff}{the anticipated effect size for average (bio)equivalence. Let \eqn{\mu_{jk}} be the population mean of the normal observations in the \eqn{k}th sequence at the \eqn{j}th period. Then, the effect size \code{diff} is \eqn{\frac{1}{2}(\mu_{11} - \mu_{21}) - \frac{1}{2}(\mu_{12} - \mu_{22})}.}

\item{sigma1}{the anticipated standard deviation for intra-subject comparison in the first sequence (group 1). That is, \eqn{\sigma_1} is the standard deviation of all \eqn{y_{i11} - y_{i21}} observations, where \eqn{i} denotes the \eqn{i}th subject in the 1st sequence.}

\item{sigma2}{the anticipated standard deviation for intra-subject comparison in the second sequence (group 2). That is, \eqn{\sigma_2} is the standard deviation of all \eqn{y_{i12} - y_{i22}} observations, where \eqn{i} denotes the \eqn{i}th subject in the 2nd sequence.}

\item{deltaL}{the lower bound for the interval of equivalence (can be \code{-Inf} for noninferiority test).}

\item{deltaU}{the upper bound for the interval of equivalence (can be \code{Inf} for noninferiority test).}

\item{alpha}{the significance level \eqn{\alpha \in (0,1)}.}

\item{targetPower}{the desired statistical power of the equivalence or noninferiority test, must be a single number between 0 and 1 (exclusive). Exactly one of the following pairs must be specified: \code{targetPower} and \code{q} or \code{n1} and \code{n2}. Specify both \code{targetPower} and \code{q} to find sample sizes \code{n1} and \code{n2} that yield desired power such that \eqn{n_{2} \approx q n_{1}}.}

\item{q}{the multiplicative constant for imbalanced two-group sample size determination (\eqn{n_{2} = q n_{1}}). The default value is 1.}

\item{n1}{the sample size for group 1. This is the number of subjects assigned to the first sequence and must be a single integer such that \eqn{n_{1} \ge 2}. Exactly one of the following pairs must be specified: \code{targetPower} and \code{q} or \code{n1} and \code{n2}. Specify both \code{n1} and \code{n2} to estimate statistical power for these sample sizes.}

\item{n2}{the sample size for group 2. This is the number of subjects assigned to the first sequence and must be a single integer such that \eqn{n_{2} \ge 2}.}

\item{plot}{a logical variable indicating whether to return a plot of the power curve. If \code{n1} and \code{n2} are specified instead of \code{q} and \code{targetPower}, this variable is automatically set to \code{FALSE}. If you wish to approximate many power curves, suppressing the plots will expedite this process.}

\item{seed}{if provided, a single positive integer is used to ensure reproducibility when randomizing the Sobol' sequence via \code{sobol()} in the \code{qrng} package (Hofert and Lemieux, 2020).}

\item{sobol}{one of the following integers: \eqn{s \in \{0, 1, 2, 3, 4 \}}. When approximating the power curve using \code{targetPower} and \code{q}, \eqn{2^{s + 10}} points are generated from the Sobol' sequence. When estimating power for given sample sizes \code{n1} and \code{n2}, \eqn{2^{s + 16}} points are generated from the Sobol' sequence. The default setting is \eqn{s = 0}, which ensures that each function call should take less than two seconds. As \eqn{s} increases, the sample size calculation is less sensitive to simulation variability but takes longer to complete. However, all function calls should still take less than 30 seconds when \eqn{s = 4}.}
}
\value{
The sample sizes or power estimate are returned as a list with supplementary information. If \code{targetPower} and \code{q} are specified to find sample sizes \code{n1} and \code{n2}, a plot of the approximated power curve will also appear in the plot pane if \code{plot = TRUE}. To confirm the sample size recommendation, power will be approximated at sample sizes \code{n1 - 1} and \code{max(2, round(q(n1 - 1)))}. This power estimate should be \emph{less} than \code{targetPower}; if not, the choice for \code{q} may be suboptimal in that it may be possible to collect fewer observations in one of the two groups and still achieve the desired study power. To find a sample size that corresponds to a different \code{targetPower}, save this function's output to an object and use the \code{UpdateTargetPower()} function.
}
\description{
Approximates the power of equivalence and noninferiority tests for two-sequence, two-period crossover designs with unequal variances. One can either find sample sizes that achieve desired statistical power or estimate the power for given sample sizes. More information about the notation and terminology for \code{diff}, \code{sigma1}, and \code{sigma2} is provided in Chapter 10 of Chow et al. (2008).
}
\examples{
# specify targetPower and q to obtain sample sizes n1 and n2
DesignCrossover2x2Unequal(diff = 0.05, sigma1 = 0.4, sigma2 = 0.3, deltaL = -0.223,
deltaU = 0.223, targetPower = 0.8, q = 1, alpha = 0.05, plot = TRUE, seed = 1, sobol = 0)
# specify n1 and n2 to estimate power for this design
DesignCrossover2x2Unequal(diff = 0.05, sigma1 = 0.4, sigma2 = 0.3, deltaL = -0.223,
deltaU = 0.223, n1 = 18, n2 = 18, alpha = 0.05, plot = FALSE, seed = 1, sobol = 0)

}
\references{
Chow, S. C., J. Shao, and H. Wang. (2008). \emph{Sample size calculations in clinical research}. Chapman and Hall/CRC.
}
